Fix ListManager list clearing
When a list was cleared (aka length set to 0) then this was not implemented correctly.

PS: While adjusting the tests I also found other cases that do not work. Added tests but commented them out because I was not really able to find a way to solve it - seems to have something to do with the managed values.