Error happened again ... fix works again
Sorry .. I do not fully understand why and wen this happens and whats the solution. In fact when it happens then when we fail to add a value because of an error this happens as a followup error. Take it out and the latest added chip tests produce it when a validation error happens