Add Note for TlvEnum
In fact we also could type the enum length according to specs but we always use UInt 32 for now. This works for all enums existing currently and for Tlv encoding only "UInt" is important, so ok for now. But add to remember that we trick a bit here. We could add PR to adjust that and also generate correct length types if wanted.