Ack last message when sending StandaloneAck
A refactoring some time ago about sending Acks for duplicates introduced a way to send a Standalone Acck and provide a special ack-number. With this we can now remove a special case and allow sending proper "normal" Standalone-Acks which acks the current pending ack-number.
And also use this for incoming DataReports with suppressresponse set to true.