Enhance Attribute Expansion, add WildcardPath
This adds the handling for the WildcardPaths to attribute expansion logic. We use the Matter model for some attributes and also prepare some "static" datastructures for that to e.g. get all diagnostic clusters ony once and so optimize the filtering.

Also the cluster Model class got enhanced for this for convenient access.