Add pure JS crypto (intermediate)
This uses crypto-browserify with a patched browserify-aes library that supports aes-ccm as we need using npm override. The tests are copied from CryptoNode and validate it basically working.

I decided for an own package for now because crypto-browserify is using a lot of outdated dependencies, so these messages should not infect matter.js main package for now and can be used when needed.