Add validation for valid paths for read/write/invoke (#961)
* Add validation for valid paths for read/write/invoke

Respect the specification definition which path combinations are allowed and which not and return InvalidAction if not allowed.

* Adjust tests to new logic and checks.

es coped check code ... the new 1.3 PR will fix that

* Pin chip tests

Pin chip testing for now to last commit before Data Model increase for 1.4 which breaks BINFO_2_1 test. Long term solution we find later

* Pin chip tool try #2

* Pin chip tool try #3

* Pin chip tool try #3

* Revert exe pinning because no tag, patch tests