Refactor Message size handling
In ordert to be more generic I decided to refactor the message size logic.

In fact the "Channel" (UDP, BLE, later TCP) best knows the message size limitation it has. So with substracting protocol overheads we can build messages of the ideal maximum size. This also removes the hard coded max Matter message size of 1024 bytes which was et from times before specs were published.
This new maxPayloadSize introduced in Channels and MessageExchange is now also used when chunking messages.