Introduces StubCommissioner
This class is a minimalistic commissioning controller that allows to just execute the initial commissioning process without completion. For this it gets initialized with the Root certificate data and the fabric data from a Controller fabric that already exists. A callback is called when operational  discovery and completing the commissioning is needed.