PASE Commissioner (#982)
* getter change and code simplify

* Enhance Commissioner

Enhance the "doOperationalDeviceConnection" callback with the option to notify the commissioner that the completion of the commissioning process was done via other ways (e.g. from a server differently from the commissioner instance).

* Enhance/Streamline CommissioningController

The commisioning options got an optional callback to allow custom commissioning completion ways.
Additionally streamline the interface (breaking) to allow just to commission to a device without a connection afterwards.

* Enhance RootCertificateManager init

And allow the class to be initialized with data and without storage too.

* Split out Env options from Controller to own type

* Move NatterController to options constructor

* Adjust Controller to generally work also without IP network

... but the standard "create" is still requiring it, so external interface stays unchanged

* No custom commissioning-complete for standard controller

... because makes no sense for now.

* Introduces StubCommissioner

This class is a minimalistic commissioning controller that allows to just execute the initial commissioning process without completion. For this it gets initialized with the Root certificate data and the fabric data from a Controller fabric that already exists. A callback is called when operational  discovery and completing the commissioning is needed.

* Allows custom commissioning completion

* Better name PASECommissioner

* Changelog

* Changelog

* linter fix

* exposes data for initialization of PaseCommissioner