Freeze operational schema
When we use models operationally as schema we treat them as immutable.  We now formalize this by freezing constituent
objects.  This results in an error (assuming strict mode) for operations that would otherwise change the schema.

Includes tests for ConcentrationMeasurement cluster and a related fix for feature map configuration for derived
clusters.