Exchange optimizations (#1168)
* Do not create exchange for just a standalone ack

... because most likely the standalone ack is from a former unacknowledged message

* Do not send failure for RetryExceeded error

When we send a response to the peer and do not receive the ack we throw internally an RetransmissionLimitReachedError. In this case we do not need to send another status (most likely we also can not because last message is still unacknowledged) and we did anything we could, so ignore error. Additionally, irrelevant for us because we are just the server,

* Adjust some loglevels