Versioning support
Modifies package.json files in git to specify their version as "git" and workspace dependency versions as "*".

Adds "version.txt" in the root of the repository to designate the latest release of the repository.

Adds a new "version" tool to matter.js-tools.  It can update the version in version.txt, apply a version to workspace package.json files and create a tag for the current version.