Fix subscription error handling
In a refactoring we unexpectedly changed behavior. The catch in sendUpdateMessage was supposed to only handle StatusResponseErrors but we now handled all errors to close the subscription on one error. This is now adjusted back to as it was before. Additionally, a catch is added to the timer based execution of sendUpdate to prevent crashes.