0.10.2 fixes (#1191)
* Fix channel cleanup

In a refactoring we introduced a regression. it happened that a channel was already removed and the code was removing then the newest added channel again which is wrong clean this up.

* Fix subscription error handling

In a refactoring we unexpectedly changed behavior. The catch in sendUpdateMessage was supposed to only handle StatusResponseErrors but we now handled all errors to close the subscription on one error. This is now adjusted back to as it was before. Additionally, a catch is added to the timer based execution of sendUpdate to prevent crashes.

* Changelog

* Also execute tests on 0.10.x branch