Add Data enrichment for DataVersionFiltered interactions
When DataVersionFilters were used to keep the response as small as possible allow to enrich the returned data as if they would have been returned by the device. So upper layer code can work generically.
Also adds methods to get a list of cached clusterVersions filtered for endpoints or clusters to use when setting the dataversion filters