Controller Refactoring and Optimizations (#1255)
* Add PromiseQueue

This adds a generic Promise Queue that executes promises with a configured concurrency and delay after each  execution before catching the next entry.

* Fix typo

* Limit Subscription requests to message size

When a device with many endpoints and clusters (like a huge bridge) is subscribed with the use of DataVersionFilters it can happen that the request exceeds the allowed message size. This change makes sure that the payload size is respected and dataVersion Filters are cutted out if needed.

* Allow to execute Interactions queued

The InteractionClient low level API gets the option now to execute any command in a queued way.
This change also incorporates moving variable definitions to the place where they are needed to lower a bit needed stack size.

* Allow to execute Interactions queued

The InteractionClient low level API gets the option now to execute any command in a queued way.
This change also incorporates moving variable definitions to the place where they are needed to lower a bit needed stack size.

* Move Subscription DataCache to PeerSet

The Subscription cache was on InteractionClient before, so very temporary. This is now moved to the PeerSet and hold there for each Node. This allows the reuse of the cached data much better.

* Clear subscription cache on non-resumed sessions

When the session could not be resumed the device usually got an OTA update or any other reason caused the removal of all resumption records, so consider cache as invalid and remove.

* Fix: Resumption for Insecure sessions should not crash

During PASE/CASE and Insecure Session establishment we have no associated Fabric and also this announcement makes no sense. Skip it

* Add Data enrichment for DataVersionFiltered interactions

When DataVersionFilters were used to keep the response as small as possible allow to enrich the returned data as if they would have been returned by the device. So upper layer code can work generically.
Also adds methods to get a list of cached clusterVersions filtered for endpoints or clusters to use when setting the dataversion filters

* Enhance some debug logging

* Enhance Attribute Listener for Subscriptions

... to also return an info if the value has changed and the old Value. Both is undefined if no former value was known

* Add CommissionedNodeDeviceData

... to keep important meta data about the node. The data are fetched on first connect and updated with each connect and determine behavior for connection concurrently and subscription limits.

* Move+Enhance Meta Data Enrichment to PairedNode

... and remove from CommissioningController.
Paired Node is emitting an "initialized" event which is used to transport the data.
Beside the basic Information details we now also collect data about connection types (Ethernet, Thread, Wifi), Thread node type, ICD and Power source.

* Enhance Subscription parameters

We use the newly collected information to better determine the subscription parameters to use.
min/max Intervals are determined. When device is thread based we use the interaction queue. If we know Dataversion or Events then we also limit responses and let data be enriched.

* Enrich Meta data if needed before subscription

If the meta data are not existent or outdated query the data from remote before subscribing. Alternatively update with cached data (yes lead to a big "kind of duplicate code" but I think worth the effort.
The removed PartLists change handler is restored otherwise in next commit.

* Reinit objects after initial init

... no need to do anything new if we can refresh. And generalize reading all Attributes

* Refactor getInteractionClient

* Enhance structure update detection

Use some more details to detect if we need to update the structure

* Expose Data

* Deprecate stateInformation callback and replace by event

* Lazy Start of Node Init on create

When creating the node we laziely start the initialization and catch potential errors. In fact the internal state engine makes sure right now that a discovery and potential reconnects happen,so it is the same as with a timer of 0ms. But lets talk if you have better ideas ...

* Rename connectedNodes to initializedNodes

... because this is the new reality because we directly return the node instance without waiting for connection to have happened

* Enhance shell

* Replace callbacks by Events

... and deprecate the callbacks

* Adjust PairedNode initialization

* Adjust concurrency to 4

* Make Linter happy

* Make Linter happy

* Finalize initialization logic and events

* Rename getConnectedNode to getPairedNode

... because the node is maybe not connected already ... Breaking

* Tweak Tests

* Address review feedback

Rethrow errors and optimization

* Address review feedback

optimize logging

* Address review feedback

Yes we need to convert this into a proper storage, but I would like to do this in a follow up PR

* Move DeviceInformation into own class

* Finalizations, Tests and changelog

* changelog