Refactor ExchangeProvider
ExchangeProvider is now an abstract class with two child classes.

One is a Dedicated version where the used channel is provided on construction and only this is used. This is mainly used by Commissioning.

The Reconnectable exchange provider now always uses the latest known channel instead of a static one. The reconnectChannel callback still exists but is only used when no channel exists. It also contains an event to inform other about a change of a channel.

Additionally, the ExchangeProvider class got moved to an own file.