Validate subscription IDs on incoming datareports
Till now we accepted all datareports and just checked at the end if they are for a valid subscription id. This change adjusts this as needed to already decline invalid subscription data on first package to save bandwidth.

Additionally, we do not need to wait for the last Success Status to be sent or messenger to be closed before processing the data, so we ignore errors there but log them.

But still try to close the messenger as soon as possible to prevent resubmissions because of missing ack's.