ControllerStore rework
I know that this part is "controverse" because ClientNode might not really use this but it is fine if this stays within the "legacy world".

ControllerStore and LegacyControllerstore hide the differences between the two versions and "clientNodeSTore" factory method generates a storage context for the storage for one client node. I know that this will be different in new API later but thats fine because this can be build up again.

This class combines several classes added into MatterController by one.