Introduce "PeerNodeStore" interface
This interface described the data interface to access cached attribute values for low level usage in interaction client. The idea is to have data structures "the same" as the ones received from the wire to easiely enhance data and such for developer convenience.

This class replaces the static structure in PeerSet.