Remove $Changed from logical flow of emitter
Previously we were providing the ActionContext of the emitter to $Changed handlers.  The transaction was destroyed in this case so was not really useful.  Now we instead just provide metadata about the subject that triggered the change.

Also, the emitter was awaiting $Changed handlers if they were async.  There are advantages and disadvantages to doing this:

* Advantage is that the logical flow is more deterministic and we track the promise as a normal part of node activity
* The disadvantage is that the emitter blocks until an independent process completes

This commit makes it so $Changed handlers operate independently of the emitter.  Tests pass but effects will be subtle so will need to keep an eye on this.