Ignore persisted values when features change (#1399)
* Ignore persisted values when features change

Affects individual behaviors on endpoints with persisted values.  Storage must be tagged with features, so storage
created in previous versions of code will not benefit from this enhancement.

Fixes #1161

* Ignore persisted values when features change

Affects individual behaviors on endpoints with persisted values.  Storage must be tagged with features, so storage
created in previous versions of code will not benefit from this enhancement.

Fixes #1161

* Fix from PR feedback