Further dynamic endpoint adjustments (#1430)
* Always add assigned numbers to allocated

... this basically ensures consistency if "nextNumber" entry would be broken in storage to not reuse any numbers

* Manage pre/allocated Numbers on close/erase

This change makes sure the pre/allocatedNumbers in the EndpointStoreService are managed when a node is erased or just closed.
When erased the number will be removed from allocated numbers to be reusable also within the process because storage is also removed. On close the allocated number is re-set to be preallocated to allow adding the same endpoint again.

* tests

* Docs fix